Reorganize execution_agent.hpp underneath execution/
