+++ agency/bulk_invoke.hpp
@@ -5,7 +5,7 @@
 #include <agency/detail/type_traits.hpp>
 #include <agency/detail/integer_sequence.hpp>
 #include <agency/detail/bulk_functions/is_bulk_call_possible_via_execution_policy.hpp>
-#include <agency/execution_agent.hpp>
+#include <agency/execution/execution_agent.hpp>
 
 namespace agency
 {
