+++ agency/execution.hpp
@@ -3,4 +3,5 @@
 #include <agency/detail/config.hpp>
 #include <agency/execution/execution_categories.hpp>
 #include <agency/execution/executor.hpp>
+#include <agency/execution/execution_agent.hpp>
